requirements.txt
python-docx>=1.1.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
pytest>=8.0.0
